Избавляемся от лишнего класса PrintPartValues
