+++ src/ru/vachok/GetWeather.java
@@ -58,5 +58,55 @@ public class GetWeather {
 
 
     }
+
+    static int printPartValues( Elements values , int index) {
+        int iterationCount = 4;
+        if (index == 0) {
+            // Проверка третьего элемента на соответствие, для корректного вывода строк
+            Element valueLn = values.get(3);
+            boolean isMorning = valueLn.text().contains("Утро");
+            boolean isDay = valueLn.text().contains("День");
+            // Если элемент массива valueLn номер 3
+            // содержит Утро - выводим только 3 первые строчки сегодняшнего дня
+            if (isMorning) {
+                iterationCount = 3;
+            }
+            if (isDay) {
+                iterationCount = 2;
+            }
+            for (int i = 0; i < iterationCount; i++) {
+                Element valueLine = values.get(index + i);
+                for (Element td : valueLine.select("td")) {
+                    System.out.print(td.text() + "    ");
+                }
+                System.out.println();
+            }
+        } else {
+            for (int i = 0; i < iterationCount; i++) {
+                // Забрать элемент массива values, по-индексу
+                // Присвоить этому имя valueLine
+                Element valueLine = values.get(index + i);
+                // Выделить из valueLine всё, что td
+                for (Element td : valueLine.select("td")) {
+                    System.out.print(td.text() + "    ");
+                }
+                System.out.println();
+            }
+        }
+        return iterationCount;
+    }
+
+    static Element printValues() {
+        Element tableWTH = null;
+        try {
+            tableWTH = GetWeather.getPage().select("table[class=wt]").first();
+        } catch (MalformedURLException e) {
+            e.printStackTrace();
+        }
+//         Elements val = tableWTH.select("")
+
+        return tableWTH;
+    }
+
 }
