+++ src/ru/vachok/Main.java
@@ -5,7 +5,7 @@ import org.jsoup.nodes.Element;
 public class Main {
 
     public static void main( String[] args ) throws Exception {
-        Element values = PrintPartValues.printValues();
+        Element values = GetWeather.printValues();
         String date = GetWeather.date();
 //        System.out.println("Погода в Санкт-Петербурге. По данным сайта pogoda.spb.ru");
         System.out.println(values);
